+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Completed ja-JP localization (#116)
+- Exposed `ObjectRegistryScope` for testing (#117)
 
 ### Fixed
 - NDMF native strings don't fall back to en-US properly (#115)